+++ src/module/workflow.ts
@@ -466,6 +466,10 @@ export class Workflow {
         break;
 
       case WORKFLOWSTATES.WAITFORATTACKROLL:
+        if (this.item.data.type === "tool" && autoFastForwardAbilityRolls) {
+          this.item.rollToolCheck({ fastForward: this.rollOptions.fastForward, advantage: this.rollOptions.advantage, disadvantage: this.rollOptions.disadvantage })
+          return this.next(WORKFLOWSTATES.WAITFORDAMAGEROLL);
+        }
         if (!this.item.hasAttack) {
           return this.next(WORKFLOWSTATES.WAITFORDAMAGEROLL);
         }
@@ -703,7 +707,7 @@ export class Workflow {
         warn('Inside workflow.rollFINISHED');
         const hasConcentration = this.item?.data.data.components?.concentration || this.item?.data.data.activation?.condition?.includes("Concentration");
         const checkConcentration = configSettings.concentrationAutomation; // installedModules.get("combat-utility-belt") && configSettings.concentrationAutomation;
-        if (hasConcentration && checkConcentration && this.applicationTargets) {
+        if (hasConcentration && checkConcentration) {
           let targets = [];
           for (let hit of this.applicationTargets) 
             targets.push({tokenUuid: hit.document.uuid, actorUuid: hit.actor.uuid});
@@ -1172,7 +1176,7 @@ export class Workflow {
         if (!!!game.dice3d?.messageHookDisabled) this.hideTags = [".midi-qol-saves-display"];
         switch (this.workflowType) {
           case "BetterRollsWorkflow":
-            const html = `<div data-item-id="${this.item._id}"></div><div class="midi-qol-saves-display">${saveHTML}${saveContent}</div><footer class="card-footer">`
+            const html = `<div data-item-id="${this.item.id}"></div><div class="midi-qol-saves-display">${saveHTML}${saveContent}</div><footer class="card-footer">`
             const roll = this.roll;
             await roll.entries.push({ type: "raw", html, source: "midi" });
             return;
@@ -1180,8 +1184,7 @@ export class Workflow {
         case "Workflow":
         case "TrapWorkflow":
             searchString =  /<div class="midi-qol-saves-display">[\s\S]*?<div class="end-midi-qol-saves-display">/;
-            // replaceString = `<div data-item-id="${this.item._id}"></div><div class="midi-qol-saves-display"><div class="midi-qol-nobox midi-qol-bigger-text">${saveFlavor}</div>${saveContent}</div>`
-            replaceString = `<div class="midi-qol-saves-display"><div data-item-id="${this.item._id}">${saveHTML}${saveContent}</div><div class="end-midi-qol-saves-display">`
+            replaceString = `<div class="midi-qol-saves-display"><div data-item-id="${this.item.id}">${saveHTML}${saveContent}</div><div class="end-midi-qol-saves-display">`
             content = content.replace(searchString, replaceString);
             await chatMessage.update({
               content, 
@@ -1201,7 +1204,7 @@ export class Workflow {
       chatData = {
         user: gmUser._id,
         speaker,
-        content: `<div data-item-id="${this.item._id}"></div> ${saveContent}`,
+        content: `<div data-item-id="${this.item.id}"></div> ${saveContent}`,
         flavor: `<h4>${this.saveDisplayFlavor}</h4>`, 
         type: CONST.CHAT_MESSAGE_TYPES.OTHER,
         flags: { "midi-qol": {type: MESSAGETYPES.SAVES, waitForDiceSoNice: waitForDSN}}